dev-requirements.txt
matplotlib
nltk
pandas
pick
seaborn
argparse
jupyter
poetry
pytest
scikit-learn
sqlalchemy
psycopg2
python-dotenv
streamlit
